backend/requirements.txt
# -------------------------
# Core Backend Framework
# -------------------------
fastapi>=0.109.0
uvicorn[standard]
pydantic-settings>=2.0.0

# -------------------------
# Auth & Security
# -------------------------
python-jose[cryptography]
bcrypt>=4.0.0
pydantic>=2.0.0
python-multipart

# -------------------------
# Database (Auth)
# -------------------------
sqlmodel>=0.0.14
asyncpg>=0.29.0
alembic>=1.13.0

# -------------------------
# Agent Orchestration
# -------------------------
langgraph
langchain-core

# -------------------------
# Biomedical NLP
# -------------------------
transformers
torch
sentencepiece

# -------------------------
# Knowledge Graph
# -------------------------
neo4j

# -------------------------
# Audit Logging
# -------------------------
cassandra-driver

# -------------------------
# Cryptography & Hashing
# -------------------------
cryptography

# -------------------------
# PDF Generation
# -------------------------
reportlab

# -------------------------
# Utilities
# -------------------------
python-dotenv
pyyaml
rich
loguru

# -------------------------
# Testing
# -------------------------
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-cov>=4.0.0
httpx
